Extracting `AggTypes` class.
